tg-bot-py/requirements.txt
python-telegram-bot==20.7
requests==2.31.0
PyYAML==6.0.1